requirements.txt
gdown==4.7.1
wandb==0.15.2
omegaconf==2.1.2
tqdm
Pillow
pytorch_fid==0.3.0
piq
fsspec
networkx
einops
pyfacer
scipy==1.10.1